fixed some coordinate system and performance issues

+++ graphics/src/core/vulkan/swap_chain.cpp
@@ -81,7 +81,7 @@ namespace undicht {
             /// @return the id of the newly acquired image, -1 if no image could be acquired (swapchain needs to be recreated)
 
             uint32_t next_image;
-            VkResult result = vkAcquireNextImageKHR(_device_handle, _swap_chain, UINT64_MAX, signal_sem, signal_fen, &next_image);
+            VkResult result = vkAcquireNextImageKHR(_device_handle, _swap_chain, UINT32_MAX, signal_sem, signal_fen, &next_image);
 
             if(result != VK_SUCCESS) {
                 UND_WARNING << "swapchain is out of date \n";
